move NewFromTemplateButton back in to TemplateList
