+++ app/templates/TemplateList.tsx
@@ -1,7 +1,10 @@
+"use client";
+
 import { ButtonPrimary } from "components/Buttons";
 import Image from "next/image";
 import Link from "next/link";
-import { NewFromTemplateButton } from "./NewFromTemplateButton";
+import { createNewLeafletFromTemplate } from "actions/createNewLeafletFromTemplate";
+import { AddTiny } from "components/Icons";
 
 export function LeafletTemplate(props: {
   title: string;
@@ -134,5 +137,20 @@ export function TemplateListExamples() {
         templateID="23d8a4ec-b2f6-438a-933d-726d2188974d"
       />
     </TemplateList>
+  );
+}
+
+export function NewFromTemplateButton(props: { templateID: string }) {
+  return (
+    <ButtonPrimary
+      className="!w-fit !border-2 !border-white hover:!outline-none hover:scale-105 hover:-rotate-2 transition-all"
+      onClick={async () => {
+        let id = await createNewLeafletFromTemplate(props.templateID, false);
+        window.open(`/${id}`, "_blank");
+      }}
+    >
+      Create
+      <AddTiny />
+    </ButtonPrimary>
   );
 }
